app/requirements.txt
fastapi
uvicorn[standard]
pyyaml
openai
httpx
aiofiles
requests